refactor: [EL-4914] restructure interactive tour context and provider
