--- src/[fsd]/app/providers/InteractiveTourContext.js
@@ -1,5 +1,0 @@
-import { createContext, useContext } from 'react';
-
-export const InteractiveTourContext = createContext(null);
-
-export const useInteractiveTour = () => useContext(InteractiveTourContext);
