+++ src/[fsd]/app/providers/index.js
@@ -4,4 +4,3 @@ export * from './AgentsStudioContext';
 export * from './ToolkitSocketContext.jsx';
 export * from './InstructionsInputRefContext.jsx';
 export * from './InteractiveTourProvider';
-export * from './InteractiveTourContext';
