+++ src/[fsd]/features/interactive-tours/index.js
@@ -1,5 +1,5 @@
 export { CARD_WIDTH_PX } from './lib/constants';
-export { isTourCompleted, useProposeTour, useTourFromUrl } from './lib/hooks';
+export { isTourCompleted, useInteractiveTourController, useProposeTour, useTourFromUrl } from './lib/hooks';
 export { AGENT_TOUR_ID } from './lib/agentTour';
 export { CHAT_TOUR_ID } from './lib/chatTour';
 export { SIDEBAR_TOUR_ID } from './lib/sidebarTour';
