+++ src/[fsd]/features/interactive-tours/lib/helpers/index.js
@@ -1,0 +1,7 @@
+export {
+  initialState,
+  lsCompletedKey,
+  lsPromptKey,
+  TOUR_COMPLETION_CONFIGS,
+  tourReducer,
+} from './interactiveTour.helpers';
