+++ src/[fsd]/features/interactive-tours/lib/hooks/index.js
@@ -1,3 +1,4 @@
+export { useInteractiveTourController } from './useInteractiveTourController.hooks';
 export { isTourCompleted, useProposeTour } from './useProposeTour.hooks';
 export { useTourCardPosition } from './useTourCardPosition.hooks';
 export { useTourFromUrl } from './useTourFromUrl.hooks';
