+++ src/[fsd]/features/interactive-tours/ui/InteractiveTourCard.jsx
@@ -4,7 +4,7 @@ import { MuiMarkdown } from 'mui-markdown';
 
 import { Box, Typography } from '@mui/material';
 
-import { useInteractiveTour } from '@/[fsd]/app/providers/InteractiveTourContext';
+import { useInteractiveTour } from '@/[fsd]/app/providers/InteractiveTourProvider';
 import { MarkdownMapping } from '@/[fsd]/shared/lib/utils/markdown.utils';
 import BaseBtn, { BUTTON_VARIANTS } from '@/[fsd]/shared/ui/button/BaseBtn';
 import { keyframes } from '@emotion/react';
