+++ src/[fsd]/features/interactive-tours/ui/TourCompleteCard.jsx
@@ -2,7 +2,7 @@ import { memo, useCallback } from 'react';
 
 import { Box, Unstable_TrapFocus as TrapFocus, Typography } from '@mui/material';
 
-import { useInteractiveTour } from '@/[fsd]/app/providers/InteractiveTourContext';
+import { useInteractiveTour } from '@/[fsd]/app/providers/InteractiveTourProvider';
 import BaseBtn, { BUTTON_VARIANTS } from '@/[fsd]/shared/ui/button/BaseBtn';
 import TutorialsSuccessIconDark from '@/assets/tutorials-success-icon-dark.svg?react';
 import TutorialsSuccessIconLight from '@/assets/tutorials-success-icon-light.svg?react';
